+++ app/src/main/java/com/hover/starter/MainActivity.java
@@ -14,8 +14,9 @@ import androidx.core.content.ContextCompat;
 
 import com.hover.sdk.api.Hover;
 import com.hover.sdk.permissions.PermissionActivity;
-import com.hover.starter.ui.main.HoverActionListAdapter;
-import com.hover.starter.ui.main.MainFragment;
+import com.hover.starter.actions.ui.actiondetail.ActionDetailActivity;
+import com.hover.starter.actions.ui.main.HoverActionListAdapter;
+import com.hover.starter.actions.ui.main.MainFragment;
 
 public class MainActivity extends AppCompatActivity implements PermissionsListener,
         HoverActionListAdapter.OnActionListItemClickListener {
@@ -65,7 +66,7 @@ public class MainActivity extends AppCompatActivity implements PermissionsListen
 
     @Override
     public void onActionListItemClick(String actionId) {
-        Intent intent = new Intent(this, ActionDetail.class);
+        Intent intent = new Intent(this, ActionDetailActivity.class);
         intent.putExtra("action_id", actionId);
         startActivity(intent);
     }
