+++ app/src/main/java/com/hover/starter/actions/data/HoverActionDao.java
@@ -1,4 +1,4 @@
-package com.hover.starter.data.actions;
+package com.hover.starter.actions.data;
 
 import androidx.lifecycle.LiveData;
 import androidx.room.Dao;
@@ -6,6 +6,8 @@ import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.Query;
 import androidx.room.Update;
+
+import com.hover.starter.actions.data.HoverAction;
 
 import java.util.List;
