+++ app/src/main/java/com/hover/starter/actions/data/HoverTransactionDao.java
@@ -1,4 +1,4 @@
-package com.hover.starter.data.transactions;
+package com.hover.starter.actions.data;
 
 import androidx.lifecycle.LiveData;
 import androidx.room.Dao;
