+++ app/src/main/java/com/hover/starter/actions/ui/main/ActionViewModel.java
@@ -1,12 +1,12 @@
-package com.hover.starter.ui.main;
+package com.hover.starter.actions.ui.main;
 
 import android.app.Application;
 
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
-import com.hover.starter.data.actions.HoverAction;
-import com.hover.starter.data.HoverRepository;
+import com.hover.starter.actions.data.HoverAction;
+import com.hover.starter.HoverRepository;
 
 import java.util.List;
