+++ app/src/main/java/com/hover/starter/actions/ui/main/HoverActionListAdapter.java
@@ -1,4 +1,4 @@
-package com.hover.starter.ui.main;
+package com.hover.starter.actions.ui.main;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -9,7 +9,7 @@ import android.widget.TextView;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.hover.starter.R;
-import com.hover.starter.data.actions.HoverAction;
+import com.hover.starter.actions.data.HoverAction;
 
 import java.util.List;
