+++ app/src/main/res/layout/action_detail_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.hover.starter.ActionDetail"
+    tools:context="com.hover.starter.actions.ui.actiondetail.ActionDetailActivity"
     tools:ignore="MergeRootFrame">
 
     <com.google.android.material.appbar.AppBarLayout
